decomp
- Read the header
- Each code is read as character and parent_id separately, to exploit the variable length of the parent_id

+++ source/decoder/decomp.h
@@ -8,17 +8,14 @@
  * ---------------------------------------------------------------------
  * 
  */
+#include <string.h>
+#include <math.h>
 #include "../common/bitio.h"
+#include "../common/util.h"
 
 typedef struct _code_t_ {
     char character;
-    uint8_t parent_id;
-    /* the last member is padded with the number of bytes required so that the
-     * total size of the structure should be a multiple of the largest alignment
-     * of any structure member:
-     * -) uint16_t -> sizeof(code_t)=4	(aligned to 2 bytes)
-     * -) uint8_t -> sizeof(code_t)=2	(aligned to 1 byte)
-     */
+    uint64_t parent_id;
 } code_t;
 
 const char *get_filename_ext(const char *filename);
